Enhancement: Split auto review tests from unit tests
