+++ Makefile
@@ -15,6 +15,7 @@ stan: vendor
 	vendor/bin/phpstan analyse --configuration=phpstan.neon --level=max src
 
 test: vendor
+	vendor/bin/phpunit --configuration=test/AutoReview/phpunit.xml
 	vendor/bin/phpunit --configuration=test/Unit/phpunit.xml
 
 vendor: composer.json composer.lock
