+++ appveyor.yml
@@ -49,4 +49,5 @@ install:
 
 test_script:
   - cd c:\projects\composer-normalize
+  - vendor/bin/phpunit --configuration=test/AutoReview/phpunit.xml
   - vendor/bin/phpunit --configuration=test/Unit/phpunit.xml
