+++ test/AutoReview/SrcCodeTest.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * @see https://github.com/localheinz/composer-normalize
  */
 
-namespace Localheinz\Composer\Normalize\Test\Unit;
+namespace Localheinz\Composer\Normalize\Test\AutoReview;
 
 use Localheinz\Test\Util\Helper;
 use PHPUnit\Framework;
@@ -19,26 +19,21 @@ use PHPUnit\Framework;
 /**
  * @internal
  */
-final class ProjectCodeTest extends Framework\TestCase
+final class SrcCodeTest extends Framework\TestCase
 {
     use Helper;
 
-    public function testProductionClassesAreAbstractOrFinal(): void
+    public function testSrcClassesAreAbstractOrFinal(): void
     {
         $this->assertClassesAreAbstractOrFinal(__DIR__ . '/../../src');
     }
 
-    public function testProductionClassesHaveTests(): void
+    public function testSrcClassesHaveTests(): void
     {
         $this->assertClassesHaveTests(
             __DIR__ . '/../../src',
             'Localheinz\\Composer\\Normalize\\',
             'Localheinz\\Composer\\Normalize\\Test\\Unit'
         );
-    }
-
-    public function testTestClassesAreAbstractOrFinal(): void
-    {
-        $this->assertClassesAreAbstractOrFinal(__DIR__ . '/..');
     }
 }
