+++ test/AutoReview/TestCodeTest.php
@@ -1,0 +1,30 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2018 Andreas Möller.
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE file that was distributed with this source code.
+ *
+ * @see https://github.com/localheinz/composer-normalize
+ */
+
+namespace Localheinz\Composer\Normalize\Test\AutoReview;
+
+use Localheinz\Test\Util\Helper;
+use PHPUnit\Framework;
+
+/**
+ * @internal
+ */
+final class TestCodeTest extends Framework\TestCase
+{
+    use Helper;
+
+    public function testTestClassesAreAbstractOrFinal(): void
+    {
+        $this->assertClassesAreAbstractOrFinal(__DIR__ . '/..');
+    }
+}
